refactor: station 패키지 분리
