--- src/main/java/subway/StationRepository.java
@@ -1,10 +1,0 @@
-package subway;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-
-    List<Station> findAllByIdIn(List<Long> ids);
-}
