+++ src/main/java/subway/station/domain/StationRepository.java
@@ -1,0 +1,14 @@
+package subway.station.domain;
+
+import java.util.List;
+
+public interface StationRepository {
+
+    List<Station> findAllByIdIn(List<Long> ids);
+
+    Station save(Station station);
+
+    List<Station> findAll();
+
+    void deleteById(Long id);
+}
