+++ src/main/java/subway/station/infra/JpaStationRepository.java
@@ -1,0 +1,9 @@
+package subway.station.infra;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import subway.station.domain.Station;
+import subway.station.domain.StationRepository;
+
+public interface JpaStationRepository extends JpaRepository<Station, Long>, StationRepository {
+
+}
